requirement.txt
absl-py==0.11.0
alembic==1.4.1
astunparse==1.6.3
attrs==20.3.0
azure-core==1.9.0
azure-storage-blob==12.6.0
bleach==3.2.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
databricks-cli==0.14.1
docutils==0.16
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
msrest==0.6.19
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pkginfo==1.7.0
pluggy==0.13.1
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.4
pyparsing==2.4.7
pytest==6.2.2
python-editor==1.0.4
querystring-parser==1.2.4
readme-renderer==28.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7
six==1.15.0
smmap==3.0.4
sqlparse==0.4.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
toml==0.10.2
tqdm==4.56.0
twine==3.3.0
typing-extensions==3.7.4.3
urllib3==1.26.3
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
